requirements.txt
Django==1.9.1
requests==2.10.0
vcrpy==1.9.0
